finish

+++ api/user/user.controller.js
@@ -1,9 +1,11 @@
 import { userService } from './user.service.js'
-
+import { bugService } from '../bug/bug.service.js'
 export async function getUsers(req, res) {
     try {
-        const users = await userService.query()
-        res.send(users)
+        _isAdmin(req, res, async () => {
+            const users = await userService.query()
+            res.send(users)
+        })
     } catch (err) {
         res.status(500).send(err)
     }
@@ -12,8 +14,10 @@ export async function getUsers(req, res) {
 export async function getUser(req, res) {
     try {
         const { userId } = req.params
-        const user = await userService.getById(userId)
-        res.send(user)
+        _isAdmin(req, res, async () => {
+            const user = await userService.getById(userId)
+            res.send(user)
+        })
     } catch (err) {
         res.status(500).send(err)
     }
@@ -21,8 +25,10 @@ export async function getUser(req, res) {
 
 export async function updateUser(req, res) {
     try {
-        const user = userService.save(req.body)
-        res.send(user)
+        _isAdmin(req, res, async () => {
+            const user = userService.save(req.body)
+            res.send(user)
+        })
     } catch (err) {
         loggerService.error('Failed to update user', err)
         res.status(400).send({ err: 'Failed to update user' })
@@ -30,12 +36,34 @@ export async function updateUser(req, res) {
 }
 
 export async function removeUser(req, res) {
+    console.log('removeUser', req.params)
+
     try {
-        const { userId } = req.params
-        const user = userService.remove(userId)
-        res.send({ msg: 'Deleted successfully' })
+        _isAdmin(req, res, async () => {
+            const hasBugs = await bugService.query({ userId: req.params.userId })
+            if (hasBugs?.bugs?.some(bug => bug?.creator?._id === req.params.userId)) {
+                res.status(403).send('User has bugs')
+                return
+            }
+            const { userId } = req.params
+            const user = userService.remove(userId)
+            res.send({ msg: 'Deleted successfully' })
+        })
     } catch (err) {
         loggerService.error('Failed to delete user', err)
         res.status(400).send({ err: 'Failed to delete user' })
+    }
+}
+
+async function _isAdmin(req, res, next) {
+    try {
+        const user = await userService.getById(req.loggedinUser._id)
+        if (user.role !== 'admin') {
+            res.status(403).send('Admin required')
+            return
+        }
+        next()
+    } catch (err) {
+        res.status(500).send(err)
     }
 }
